refactor: restructure TMDB provider into modular architecture
Split the monolithic tmdb.rs file (1,423 lines) into a clean module structure:
- base.rs: Core TmdbService with shared functionality
- models.rs: Shared data structures and type definitions
- movies.rs: Movie-specific provider implementation
- shows.rs: TV show provider implementation
- non_metadata.rs: Person and company provider implementation

This improves code organization, maintainability, and follows Rust module conventions while preserving all existing functionality.
