+++ crates/providers/src/tmdb/mod.rs
@@ -1,0 +1,13 @@
+mod base;
+pub use base::TmdbService;
+
+mod models;
+
+mod movies;
+pub use movies::TmdbMovieService;
+
+mod non_metadata;
+pub use non_metadata::NonMediaTmdbService;
+
+mod shows;
+pub use shows::TmdbShowService;
